version_airflow/requirements.txt
requests==2.32.3
google-cloud-bigquery==3.25.0a